standardize variables
Most of the page used the variable standard of caps and hyphens
but then some of the page was using <lowercase> text for what I think was meant to be
variables. I removed the angle brackets and capitalized the text.
Going forward, our house style is to include an example anytime you give syntax with a variable.
See https://docs-wiki.cfapps.io/wiki/style/documenting-commands.html